+++ Heuristic/Rules/IScoringRule.cs
@@ -6,7 +6,7 @@ namespace Kate.Heuristic.Rules
 {
     public interface IScoringRule
     {
-        decimal evaluateScore(IMap map);
+        float evaluateScore(IMap map);
     }
 }
